+++ flash/text/seq2seq/summarization/data.py
@@ -13,22 +13,38 @@
 # limitations under the License.
 from typing import Any
 
-from transformers import AutoTokenizer
+from flash.core.data.process import Postprocess
+from flash.core.utilities.imports import _TEXT_AVAILABLE
+from flash.text.seq2seq.core.data import Seq2SeqBackboneState, Seq2SeqData, Seq2SeqPreprocess
 
-from flash.data.process import Postprocess
-from flash.text.seq2seq.core.data import Seq2SeqData, Seq2SeqPreprocess
+if _TEXT_AVAILABLE:
+    from transformers import AutoTokenizer
 
 
 class SummarizationPostprocess(Postprocess):
 
-    def __init__(
-        self,
-        backbone: str = "sshleifer/tiny-mbart",
-    ):
+    def __init__(self):
         super().__init__()
 
+        if not _TEXT_AVAILABLE:
+            raise ModuleNotFoundError("Please, pip install -e '.[text]'")
+
         # TODO: Should share the backbone or tokenizer over state
-        self.tokenizer = AutoTokenizer.from_pretrained(backbone, use_fast=True)
+        self._backbone = None
+        self._tokenizer = None
+
+    @property
+    def backbone(self):
+        backbone_state = self.get_state(Seq2SeqBackboneState)
+        if backbone_state is not None:
+            return backbone_state.backbone
+
+    @property
+    def tokenizer(self):
+        if self.backbone is not None and self.backbone != self._backbone:
+            self._tokenizer = AutoTokenizer.from_pretrained(self.backbone, use_fast=True)
+            self._backbone = self.backbone
+        return self._tokenizer
 
     def uncollate(self, generated_tokens: Any) -> Any:
         pred_str = self.tokenizer.batch_decode(generated_tokens, skip_special_tokens=True)