+++ flash/vision/detection/__init__.py
@@ -1,2 +1,0 @@
-from flash.vision.detection.data import ObjectDetectionData
-from flash.vision.detection.model import ObjectDetector